emulator/config: remove version from config
