+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/Config.java
@@ -1,0 +1,4 @@
+package org.ethereum.beacon.emulator.config;
+
+/** Config interface. Doesn't require anything just marking top-level configs */
+public interface Config {}
