+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/ConfigSource.java
@@ -1,0 +1,11 @@
+package org.ethereum.beacon.emulator.config;
+
+/** Source of config {@link org.ethereum.beacon.emulator.config.Config} */
+public interface ConfigSource {
+  Type getType();
+
+  public enum Type {
+    ASIS,
+    YAML
+  }
+}
