+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/ConfigSupplier.java
@@ -1,6 +1,6 @@
 package org.ethereum.beacon.emulator.config;
 
-import org.ethereum.beacon.emulator.config.data.Config;
+import org.ethereum.beacon.emulator.config.Config;
 
 /** Uses some data to construct {@link Config} */
 public interface ConfigSupplier {
