--- emulator/config/src/main/java/org/ethereum/beacon/emulator/config/data/Config.java
@@ -1,9 +1,0 @@
-package org.ethereum.beacon.emulator.config.data;
-
-/**
- * Abstract configuration, any configuration to be used with {@link
- * org.ethereum.beacon.emulator.config.ConfigBuilder} should implement this interface
- */
-public interface Config {
-  int getVersion();
-}
