+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/main/Configuration.java
@@ -1,4 +1,4 @@
-package org.ethereum.beacon.emulator.config.data.v1;
+package org.ethereum.beacon.emulator.config.main;
 
 /** Beacon chain configuration */
 public class Configuration {
