+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/main/action/ActionRun.java
@@ -1,4 +1,4 @@
-package org.ethereum.beacon.emulator.config.data.v1.action;
+package org.ethereum.beacon.emulator.config.main.action;
 
 /** Indicates that service that belongs to to this action should be started and run */
 public class ActionRun extends Action {
