--- emulator/config/src/main/java/org/ethereum/beacon/emulator/config/type/ConfigSource.java
@@ -1,11 +1,0 @@
-package org.ethereum.beacon.emulator.config.type;
-
-/** Source of config {@link org.ethereum.beacon.emulator.config.data.Config} */
-public interface ConfigSource {
-  Type getType();
-
-  public enum Type {
-    ASIS,
-    YAML
-  }
-}
