+++ emulator/config/src/test/resources/chainSpec.yml
@@ -1,4 +1,3 @@
-version: 1
 depositContractParameters:
   DEPOSIT_CONTRACT_ADDRESS: 0x0000000000000000000000000000000000000000
   DEPOSIT_CONTRACT_TREE_DEPTH: 32
