+++ emulator/config/src/test/resources/config.yml
@@ -1,4 +1,3 @@
-version: 1
 config:
   chainSpec: file://chainSpec.json
   db: file://db
